+++ packages/sensai-doc/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.4.1 (2025-06-02)
+
+**Note:** Version bump only for package @sensai/doc
+
+# 1.4.0 (2025-05-31)
+
+### Features
+
+- **guard:** Add basic guard module ([#12](https://github.com/sensaihq/sensai/issues/12)) ([99dad14](https://github.com/sensaihq/sensai/commit/99dad14a5242497be36c73e185e47a304c5ab2f5))
+
 ## <small>1.3.13 (2025-05-26)</small>
 
 - chore(versioning): Only accept feat and fix for version bumping ([47fc9e0](https://github.com/sensaihq/sensai/commit/47fc9e0))